Task#27 Update list of commands on telegram api on bot initialization Changed name of package
