+++ telegram/src/integrationTest/java/by/mrrockka/aspect/TelegramExceptionHandlerTest.java
@@ -1,11 +1,11 @@
 package by.mrrockka.aspect;
 
+import by.mrrockka.bot.PokerClockAbsSender;
+import by.mrrockka.bot.PokerClockBot;
 import by.mrrockka.creator.ChatCreator;
 import by.mrrockka.creator.MessageCreator;
 import by.mrrockka.creator.UpdateCreator;
 import by.mrrockka.exception.BusinessException;
-import by.mrrockka.route.PokerClockAbsSender;
-import by.mrrockka.route.PokerClockBot;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
