+++ telegram/src/main/java/by/mrrockka/bot/commands/BountyTelegramCommand.java
@@ -1,6 +1,6 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
-import by.mrrockka.service.TelegramHelpService;
+import by.mrrockka.service.TelegramBountyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethodMessage;
@@ -8,13 +8,14 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 
 @Component
 @RequiredArgsConstructor
-public class HelpTelegramCommand implements TelegramCommand {
-  private static final String COMMAND = "^/help$";
-  private final TelegramHelpService telegramHelpService;
+public class BountyTelegramCommand implements TelegramCommand {
+  private static final String COMMAND = "^/bounty$";
+
+  private final TelegramBountyService telegramBountyService;
 
   @Override
   public BotApiMethodMessage process(final Update update) {
-    return telegramHelpService.calculatePayments(update);
+    return telegramBountyService.storeBounty(update);
   }
 
   @Override
