+++ telegram/src/main/java/by/mrrockka/bot/commands/CalculateTelegramCommand.java
@@ -1,4 +1,4 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
 import by.mrrockka.service.TelegramCalculationService;
 import lombok.RequiredArgsConstructor;
