+++ telegram/src/main/java/by/mrrockka/bot/commands/CashGameTelegramCommand.java
@@ -1,6 +1,6 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
-import by.mrrockka.service.TelegramPrizePoolService;
+import by.mrrockka.service.game.TelegramGameService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethodMessage;
@@ -8,14 +8,13 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 
 @Component
 @RequiredArgsConstructor
-public class PrizePoolTelegramCommand implements TelegramCommand {
-  private static final String COMMAND = "^/prizepool$";
-
-  private final TelegramPrizePoolService prizePoolService;
+public class CashGameTelegramCommand implements TelegramCommand {
+  private static final String COMMAND = "^/cash$";
+  private final TelegramGameService gameService;
 
   @Override
   public BotApiMethodMessage process(final Update update) {
-    return prizePoolService.storePrizePool(update);
+    return gameService.storeCashGame(update);
   }
 
   @Override
