+++ telegram/src/main/java/by/mrrockka/bot/commands/EntryTelegramCommand.java
@@ -1,6 +1,6 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
-import by.mrrockka.service.game.TelegramGameService;
+import by.mrrockka.service.TelegramEntryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethodMessage;
@@ -8,13 +8,14 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 
 @Component
 @RequiredArgsConstructor
-public class CashGameTelegramCommand implements TelegramCommand {
-  private static final String COMMAND = "^/cash$";
-  private final TelegramGameService gameService;
+public class EntryTelegramCommand implements TelegramCommand {
+  private static final String COMMAND = "^/entry$";
+
+  private final TelegramEntryService telegramEntryService;
 
   @Override
   public BotApiMethodMessage process(final Update update) {
-    return gameService.storeCashGame(update);
+    return telegramEntryService.storeEntry(update);
   }
 
   @Override
