+++ telegram/src/main/java/by/mrrockka/bot/commands/FinalePlacesTelegramCommand.java
@@ -1,6 +1,6 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
-import by.mrrockka.service.TelegramWithdrawalService;
+import by.mrrockka.service.TelegramFinalePlacesService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethodMessage;
@@ -8,18 +8,19 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 
 @Component
 @RequiredArgsConstructor
-public class WithdrawalTelegramCommand implements TelegramCommand {
-  private static final String COMMAND = "^/withdrawal$";
+public class FinalePlacesTelegramCommand implements TelegramCommand {
+  private static final String COMMAND = "^/finaleplaces$";
 
-  private final TelegramWithdrawalService telegramWithdrawalService;
+  private final TelegramFinalePlacesService telegramFinalePlacesService;
 
   @Override
   public BotApiMethodMessage process(final Update update) {
-    return telegramWithdrawalService.storeWithdrawal(update);
+    return telegramFinalePlacesService.storePrizePool(update);
   }
 
   @Override
   public String commandPattern() {
     return COMMAND;
   }
+
 }
