+++ telegram/src/main/java/by/mrrockka/bot/commands/PrizePoolTelegramCommand.java
@@ -1,6 +1,6 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
-import by.mrrockka.service.TelegramEntryService;
+import by.mrrockka.service.TelegramPrizePoolService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethodMessage;
@@ -8,14 +8,14 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 
 @Component
 @RequiredArgsConstructor
-public class EntryTelegramCommand implements TelegramCommand {
-  private static final String COMMAND = "^/entry$";
+public class PrizePoolTelegramCommand implements TelegramCommand {
+  private static final String COMMAND = "^/prizepool$";
 
-  private final TelegramEntryService telegramEntryService;
+  private final TelegramPrizePoolService prizePoolService;
 
   @Override
   public BotApiMethodMessage process(final Update update) {
-    return telegramEntryService.storeEntry(update);
+    return prizePoolService.storePrizePool(update);
   }
 
   @Override
