+++ telegram/src/main/java/by/mrrockka/bot/commands/TelegramCommand.java
@@ -1,4 +1,4 @@
-package by.mrrockka.route.commands;
+package by.mrrockka.bot.commands;
 
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethodMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
