+++ telegram/src/main/java/by/mrrockka/bot/properties/BotProperties.java
@@ -1,4 +1,4 @@
-package by.mrrockka.route.properties;
+package by.mrrockka.bot.properties;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
