+++ telegram/src/main/java/by/mrrockka/bot/properties/CommandDescription.java
@@ -1,4 +1,4 @@
-package by.mrrockka.route.properties;
+package by.mrrockka.bot.properties;
 
 public record CommandDescription(String description, String details) {
 }
