+++ telegram/src/main/java/by/mrrockka/service/UpdateBotCommandsService.java
@@ -1,8 +1,8 @@
 package by.mrrockka.service;
 
+import by.mrrockka.bot.PokerClockAbsSender;
+import by.mrrockka.bot.properties.CommandDescription;
 import by.mrrockka.mapper.BotCommandMapper;
-import by.mrrockka.route.PokerClockAbsSender;
-import by.mrrockka.route.properties.CommandDescription;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
